@@ -16,23 +16,23 @@
     "babel-plugin-syntax-jsx": "^6.18.0",
     "babel-plugin-transform-vue-jsx": "^3.7.0",
     "browser-assert": "^1.2.1",
-    "cm-highlight": "^0.1.1",
     "codemirror": "^5.48.4",
-    "highlight": "^0.2.4",
-    "vue": "^2.7.0",
-    "vue-cm": "^1.1.0",
-    "vue-windows": "^0.3.0"
+    "vue": "^3.2.37"
   },
   "devDependencies": {
+    "@iconify-json/carbon": "^1.1.6",
     "@sxzz/eslint-config-prettier": "^2.2.2",
-    "@sxzz/eslint-config-vue2": "^2.2.2",
+    "@sxzz/eslint-config-vue": "^2.2.2",
     "@types/babel__core": "^7.1.19",
     "@types/babel__standalone": "^7.1.4",
+    "@types/codemirror": "^5.60.5",
     "@unocss/reset": "^0.43.0",
-    "@vitejs/plugin-vue2": "^1.1.2",
+    "@vitejs/plugin-vue": "^2.3.3",
+    "@vue/runtime-core": "^3.2.37",
     "eslint": "^8.19.0",
     "sass": "^1.22.10",
     "unocss": "^0.43.0",
+    "unplugin-auto-import": "^0.9.2",
     "unplugin-vue-components": "^0.21.0",
     "vite": "^2.9.13"
   }
