@@ -1,10 +1,6 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
 import 'uno.css'
-import '@unocss/reset/normalize.css'
+import '@unocss/reset/tailwind.css'
 
-Vue.config.productionTip = false
-
-new Vue({
-  render: (h) => h(App),
-}).$mount('#app')
+createApp(App).mount('#app')
