@@ -1,4 +1,4 @@
 module.exports = {
   root: true,
-  extends: ['@sxzz/eslint-config-vue2', '@sxzz/eslint-config-prettier'],
+  extends: ['@sxzz/eslint-config-vue', '@sxzz/eslint-config-prettier'],
 }
