@@ -7,7 +7,10 @@ export {}
 
 declare module '@vue/runtime-core' {
   export interface GlobalComponents {
-    JsxEditor: typeof import('./components/JsxEditor.vue')['default']
+    AppFooter: typeof import('./components/app-footer.vue')['default']
+    AppHeader: typeof import('./components/app-header.vue')['default']
+    CodeMirror: typeof import('./components/codemirror/CodeMirror.vue')['default']
+    JsxEditor: typeof import('./components/jsx-editor.vue')['default']
   }
 
 }
